Remove all .mjs and .umd.js references
React doesn't expose an ES Module version of React.
So all packages that depend on React surface errors in apps using Webpacker@4.
Webpacker@4 has stronger opinions about `.mjs`-suffixed files and how to resolve dependencies.
So, we just removed the `.mjs` references.
Webpacker just carries on with the CommonJS "modules".

In addition that this,
I've dropped references for `.umd.js`.
The file is still prepared and available on npm and via `node_modules`, should anyone really care for it.